refavtor: edit header, favortisePage and readme.md
