+++ src/app/(main)/favorites/page.module.scss
@@ -2,7 +2,7 @@
 
 .page {
   width: 100%;
-  min-height: calc(100dvh - $header-height);
+  min-height: calc(100dvh - 66px);
 
   display: flex;
   flex-direction: column;
