--- src/shared/fonts/index.ts
@@ -1,6 +1,0 @@
-import { Onest } from 'next/font/google'
-
-export const onest = Onest({
-  subsets: ['cyrillic', 'latin'],
-  weight: ['400', '500', '600'],
-})
