+++ src/shared/ui/container/ui/ui.tsx
@@ -1,6 +1,6 @@
 import { ReactNode } from 'react'
 import clsx from 'clsx'
-import styles from './container.module.scss'
+import styles from './ui.module.scss'
 
 interface ContainerProps {
   children: ReactNode
