+++ tsconfig.json
@@ -29,9 +29,8 @@
       "@mocks/*": ["src/mocks/*"],
       "@providers/*": ["src/app/providers/*"],
       "@styles/*": ["src/shared/styles/*"],
-      // "@icons/*": ["src/shared/icons/*"],
       "@icons/*": ["src/shared/assets/icons/*"],
-      "@assets/*": ["src/shared/assets/*"]
+      "@fonts/*": ["src/shared/assets/fonts/*"],
     },
   },
   "include": [
